fix: when user address unknown do not try to load pool data (#1311)
* test: add lighthouse test workflow


* test: add lighthouse test workflow


* test: add lighthouse test workflow


* test: add lighthouse test workflow


* docs: add contribution guide


* chore: merge .gitmodules


* docs: link to discord as contact point


* docs: update README contrib section


* Update .gitpod.yml

* feat: replace coingecko API call with OHM API


* docs: update test unit section


* docs: update test unit section


* chore: fix spelling and gitpod extensions


* fix: gitpod format

* build: fix gitpod yaml

* fix: when user address uknown do not load pool data


* fix: when user address uknown do not load pool data


* fix: extract IAddress to baseTypes.ts


* fix: extract IAddress to baseTypes.ts


* chore: split constants into multiple files


* chore: merge develop


* remove unncessary if statement

* eliminate 8 RPC calls by using gOHM price hook, rather than fetch gOHM price for each pool

* update test
