+++ README.md
@@ -39,11 +39,21 @@ Open the source code and start editing!
 
 If you would like to run the frontend in a Docker image (e.g. to isolate dependencies and the nodejs version), run `yarn docker-start`.
 
-## Unit Testing
+## Testing
 
-Unit tests are co-located with source code with naming convention `*.unit.test.js`.
+We use the [React Jest](https://jestjs.io/docs/tutorial-react) test driver for unit tests, snapshot tests and e2e tests.
 
-We use the [React Jest](https://jestjs.io/docs/tutorial-react) test driver.
+To run tests in interactive mode during development:
+
+```
+yarn test
+```
+
+### Unit Testing
+
+Unit test files are co-located with the source code files that they test and follow the naming convention `*.unit.test.ts`.
+For example unit tests for `OriginalSourceFile.ts` are located in `OriginalSourceFile.unit.test.ts`.
+Valid extensions for test files are `.js` (JavaScript), `.ts` (TypeScript), `.jsx` (React JSX), `.tsx` (React TSX).
 
 To run all unit test and see a coverage report:
 
@@ -81,7 +91,7 @@ To run the tests:
 - Run the frontend, using `yarn start`
 - In another terminal, run the tests, using `yarn test:e2e`
 
-## Rinkeby Testing
+### Rinkeby Testing
 
 ### sOHM Faucet
 
@@ -90,6 +100,7 @@ To run the tests:
 - After connecting to web3, click `Write` to execute and 10 sOHM will automatically be transferred to your connected wallet.
 
 Note: The faucet is limited to one transfer per wallet every 6500 blocks (~1 day)
+Note: This faucet drips sOHM v1 tokens. If you need to test v2 token flows (sOHM, OHM, gOHM), you will first need to use the migration steps in the UI to convert from sOHM v1 to sOHM v2.
 
 ### wETH Faucet
